pyproject.toml
[project]
name = "compilador"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "antlr4-python3-runtime==4.13.2",
    "antlr4-tools==0.2.2",
    "install-jdk==1.1.0",
    "pydantic>=2.11.7",
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = [
    "ipython>=9.4.0",
]
